feat(List): Improve information provided in the list
Instead of `name` we now provide `displayName`. Additionally `profileDir` is now `profileDirPath`
and a new `profileDirName` property is available which is the one you need to pass to the browser
cli command in order to load a given profile.

BREAKING CHANGE: Properties `name` and `profileDir` are removed and replaced with `displayName`,
`profileDirName` and `profileDirPath`.